+++ src/ladb_opencutlist/ruby/tool/highlight_part_tool.rb
@@ -233,6 +233,10 @@ module Ladb::OpenCutList
 
     end
 
+    def getExtents
+      Sketchup.active_model.bounds
+    end
+
     # -- Menu --
 
     if Sketchup.version.to_i < 15