+++ CHANGELOG.md
@@ -6,10 +6,11 @@ in 3.x, 2.x, 1.x and 0.x versions.
 
 * 3.0.1 (WIP)
 
-  * Better selection feedbacks in cutlist, cutting diagrams, labels and reports
+  * Now keep dialog size and position when using SketchUp >= 2021.1
   * Save default _mass_unit_ and _currency_symbol_ to model even if it is not modified from default values
+  * Improved document title when print to PDF
+  * Improved selection warning feedbacks in cutlist, cutting diagrams, labels and reports
   * Fixed cutting diagram part list sort when using letter for part's number
-  * Now keep dialog size and position when using SketchUp >= 2021.1 
 
 * 3.0.0 (2021-11-14)